more progress on database

+++ src/pagination.rs
@@ -1,11 +1,7 @@
-// use crate::*;
-// use std::collections::HashMap;
-// use std::error::Error;
-
 use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 pub struct Pagination {
-    pub page: usize,
-    pub limit: usize,
+    pub page: i32,
+    pub limit: i32,
 }
